requirements.txt
streamlit
tensorflow
numpy
pandas
Pillow
locust
pyngrok
